Added category view and encryption in base64

+++ SystemVault.Presentation/Views/UserControls/FileView.xaml.cs
@@ -1,11 +1,12 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using System.IO;
 using System.Windows;
 using System.Windows.Controls;
 using SystemVault.BLL.DTOs.ServiceFile;
 using SystemVault.BLL.Interfaces;
 using SystemVault.DAL.Models.SearchCriteria;
 using SystemVault.Presentation.Helpers;
-using SystemVault.Presentation.Views.Windows;
+using SystemVault.Presentation.Views.Windows.ServiceFile;
 
 namespace SystemVault.Presentation.Views.UserControls;
 
@@ -15,15 +16,17 @@ public partial class FileView : UserControl
 
     private readonly IServiceFileService _serviceFileService;
     private readonly ICategoryService _categoryService;
+    private readonly ICryptoService _cryptoService;
     private ServiceFileSC _serviceFileSC;
 
-    public FileView(IServiceProvider serviceProvider, IServiceFileService serviceFileService, ICategoryService categoryService)
+    public FileView(IServiceProvider serviceProvider, IServiceFileService serviceFileService, ICategoryService categoryService, ICryptoService cryptoService)
     {
         InitializeComponent();
         _serviceProvider = serviceProvider;
 
         _serviceFileService = serviceFileService;
         _categoryService = categoryService;
+        _cryptoService = cryptoService;
         _serviceFileSC = new ServiceFileSC();
 
         cmbPageSize.SelectedIndex = 0;
@@ -94,7 +97,7 @@ public partial class FileView : UserControl
         Search(_serviceFileSC);
     }
 
-    private void FileEdit_Click(object sender, System.Windows.RoutedEventArgs e)
+    private void FileEdit_Click(object sender, RoutedEventArgs e)
     {
         var button = sender as Button;
         var selectedFile = button?.CommandParameter as ServiceFileDto;
@@ -105,12 +108,16 @@ public partial class FileView : UserControl
         addFileWindow.ShowDialog();
     }
 
-    private async void FileDelete_Click(object sender, System.Windows.RoutedEventArgs e)
+    private async void FileDelete_Click(object sender, RoutedEventArgs e)
     {
         var button = sender as Button;
         var selectedFile = button?.CommandParameter as ServiceFileDto;
 
         if (selectedFile == null) return;
+
+        var result = MessageBoxHelper.ShowYesNo("Are you sure?");
+
+        if (result != MessageBoxResult.Yes) return;
 
         await _serviceFileService.DeleteAsync(selectedFile.Id);
         await _serviceFileService.SaveChangesAsync();
@@ -118,6 +125,21 @@ public partial class FileView : UserControl
         Search(_serviceFileSC);
 
         MessageBoxHelper.ShowInfo($"{selectedFile.Name} succesfully deleted!");
+    }
+
+    private void FileExport_Click(object sender, RoutedEventArgs e)
+    {
+        var button = sender as Button;
+        var selectedFile = button?.CommandParameter as ServiceFileDto;
+
+        if (selectedFile == null) return;
+
+        var outputFilename = Path.GetFileName(selectedFile.Path);
+        var desktopPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop), outputFilename);
+
+        _cryptoService.DecryptFile(selectedFile.Path, desktopPath, "test123");
+
+        MessageBoxHelper.ShowInfo($"{selectedFile.Name} succesfully decrypted!");
     }
 
     private void CategoryComboBox_Loaded(object sender, RoutedEventArgs e)
@@ -134,3 +156,36 @@ public partial class FileView : UserControl
         }
     }
 }
+
+/*public static class AuthorizationHelper
+{
+    public static readonly DependencyProperty RequiresAuthorizationProperty =
+        DependencyProperty.RegisterAttached(
+            "RequiresAuthorization",
+            typeof(bool),
+            typeof(AuthorizationHelper),
+            new PropertyMetadata(false, OnRequiresAuthorizationChanged));
+
+    public static void SetRequiresAuthorization(DependencyObject obj, bool value) =>
+        obj.SetValue(RequiresAuthorizationProperty, value);
+
+    public static bool GetRequiresAuthorization(DependencyObject obj) =>
+        (bool)obj.GetValue(RequiresAuthorizationProperty);
+
+    private static void OnRequiresAuthorizationChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
+    {
+        if (e.NewValue is true)
+        {
+            // Add authorization logic here
+            if (!IsUserAuthorized())
+            {
+                
+                throw new UnauthorizedAccessException("User is not authorized to view this control.");
+            }
+        }
+    }
+    private static bool IsUserAuthorized()
+    {
+        return false;
+    }
+}*/
